Move profiling non-UI test into its own project.
