+++ Python/Tests/ProfilingUITests/ProfilingUITests.csproj
@@ -78,7 +78,6 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="ProfilingUITests.cs" />
-    <Compile Include="ProfilingTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="VSPackage.cs" />
   </ItemGroup>
